--- Sample/MultipleModuleSample/datasource/src/main/java/com/github/mag0716/datasource/model/Data.kt
@@ -1,7 +1,0 @@
-package com.github.mag0716.datasource.model
-
-data class Data(
-        val id: Int,
-        val title: String,
-        val thumbnail: String? = null
-)
