--- Sample/MultipleModuleSample/datasource/src/main/java/com/github/mag0716/datasource/model/Detail.kt
@@ -1,9 +1,0 @@
-package com.github.mag0716.datastore.model
-
-data class Detail(
-        val id: Int,
-        val title: String,
-        val description: String,
-        val image: String? = null,
-        val createdTime: Long
-)
