+++ Sample/MultipleModuleSample/datasource/src/main/java/com/github/mag0716/multimodulesample/datasource/DataSourceModule.kt
@@ -1,4 +1,4 @@
-package com.github.mag0716.datasource
+package com.github.mag0716.multimodulesample.datasource
 
 import com.github.mag0716.api.ApiService
