+++ Sample/MultipleModuleSample/datasource/src/main/java/com/github/mag0716/multimodulesample/datasource/converter/DetailResponseConverter.kt
@@ -1,0 +1,14 @@
+package com.github.mag0716.multimodulesample.datasource.converter
+
+import com.github.mag0716.api.response.DetailResponse
+import com.github.mag0716.multimodulesample.datastore.model.Detail
+
+internal fun DetailResponse.toDetail(): Detail {
+    return Detail(
+        this.id,
+        this.title,
+        this.description,
+        this.image,
+        this.createdTime
+    )
+}
