+++ Sample/MultipleModuleSample/datasource/src/main/java/com/github/mag0716/multimodulesample/datasource/model/Data.kt
@@ -1,0 +1,7 @@
+package com.github.mag0716.multimodulesample.datasource.model
+
+data class Data(
+    val id: Int,
+    val title: String,
+    val thumbnail: String? = null
+)
