+++ Sample/MultipleModuleSample/usecase/src/main/java/com/github/mag0716/usercase/GetDataListUseCase.kt
@@ -1,7 +1,7 @@
 package com.github.mag0716.usercase
 
-import com.github.mag0716.datasource.IDataRepository
-import com.github.mag0716.datasource.model.Data
+import com.github.mag0716.multimodulesample.datasource.IDataRepository
+import com.github.mag0716.multimodulesample.datasource.model.Data
 import timber.log.Timber
 import timber.log.debug
