+++ Sample/MultipleModuleSample/usecase/src/main/java/com/github/mag0716/usercase/IGetDataDetailUseCase.kt
@@ -1,6 +1,6 @@
 package com.github.mag0716.usercase
 
-import com.github.mag0716.datastore.model.Detail
+import com.github.mag0716.multimodulesample.datastore.model.Detail
 
 interface IGetDataDetailUseCase {
     suspend fun execute(id: Int): Detail
