+++ Sample/MultipleModuleSample/usecase/src/main/java/com/github/mag0716/usercase/IGetDataListUseCase.kt
@@ -1,6 +1,6 @@
 package com.github.mag0716.usercase
 
-import com.github.mag0716.datasource.model.Data
+import com.github.mag0716.multimodulesample.datasource.model.Data
 
 interface IGetDataListUseCase {
     suspend fun execute(): List<Data>
