+++ Sample/MultipleModuleSample/usecase/src/main/java/com/github/mag0716/usercase/UseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.github.mag0716.usercase
 
-import com.github.mag0716.datasource.IDataRepository
+import com.github.mag0716.multimodulesample.datasource.IDataRepository
 
 fun provideGetDataListUseCase(dataRepository: IDataRepository): IGetDataListUseCase {
     return GetDataListUseCase(dataRepository)
